+++ mod/view/fragment/settings_modal.php
@@ -19,8 +19,8 @@
  * --------------------------------------------------------------------
  * ZnetDK 4 Mobile User sessions module view fragment
  * 
- * File version: 1.0
- * Last update: 04/24/2025
+ * File version: 1.1
+ * Last update: 06/15/2025
  */
 ?>
 <div id="z4m-user-sessions-settings-modal" class="w3-modal">
@@ -75,6 +75,11 @@ foreach ($znetdkSettings as $zdkParamName => $anchor) : ?>
                             <th class="w3-monospace"><?php echo is_bool(constant($zdkParamName)) ? (constant($zdkParamName) ? 'true' : 'false') : constant($zdkParamName); ?></th>
                         </tr>
 <?php endforeach; ?>
+                        <tr>
+                            <td class="param">MOD_Z4M_USERSESSIONS_APPLICATION_URI</td>
+                            <td class="w3-monospace">-</td>
+                            <th class="w3-monospace"><?php echo MOD_Z4M_USERSESSIONS_APPLICATION_URI === NULL ? 'null' : "'" . MOD_Z4M_USERSESSIONS_APPLICATION_URI . "'"; ?></th>
+                        </tr>
                     </tbody>
                 </table>
             </div>